Split OpenJij adapter package by responsibility
